split off parser errors
